Added database manager, and setup basic Reminder class
Implemented an SQlite database manager, which is where our future data will be stored. Currently, the only table that can exist is the 'reminders' table.

As of right now, reminders are not actually linked to this, but every time a reminder is created via command, it gets stored inside of the database.

The Reminder class has various methods to manipulate the table, providing CRUD operations for everything you'd need.

This is to allow us to keep track of reminders and other information out-of-memory, meaning when the bot is closed, the data will be retained.