Added To-Do and Testing Option Displays
Made a basic To-Do list of stuff I was just thinking of. As well trying a couple different potential ways I could implement something along the lines of autocomplete, or even just listing a bunch of options. (thinking in the future I could use the fuzzywuzzy library in conjunction).

Attempted using Choices object but it couldn't be converted to json so that was causing a problem.

Trying to see if I can get OptionChoice to work in disnake(still experimenting)

P.S this is more for me so I can at least look back and see I was doing something today. Sorry for the long commit